+++ lib/src/geo_point.dart
@@ -1,6 +1,9 @@
-class GeoPoint{
-      final double longitude;
-      final double latitude;
+class GeoPoint {
+  final double longitude;
+  final double latitude;
 
-      GeoPoint({this.latitude,this.longitude});
+  GeoPoint({
+    this.latitude,
+    this.longitude,
+  });
 }