+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.1.0 [alpha-version]:
+
+* contain basic for android
+* trakcing,set position,current position
+
 ## 0.0.1
 
 * TODO: Describe initial release.